* [ASTERIXDB-3667][OTH] Log counts of currently running jobs based on kind (MB-69127)
 * [ASTERIXDB-3662][EXT]: Add external stats tracker (MB-68345)
 * [NO ISSUE][COMP] Fix for failing q10 of ch2++ (MB-69019)
 * [ASTERIXDB-3649][*DB] Improve async request API (MB-60882)
 * [ASTERIXDB-3649][*DB] Async request API backward compatibility fixes (MB-60882)
 * [NO ISSUE][EXT]: ensure CSV header is written in all files (MB-69147)
 * [ASTERIXDB-3664]: Simplify function name handling in EntityDetails (MB-68413)

+++ hyracks-fullstack/hyracks/hyracks-api/src/main/java/org/apache/hyracks/api/result/IResultPartitionManager.java
@@ -30,7 +30,7 @@ public interface IResultPartitionManager extends IResultManager {
     void registerResultPartitionLocation(JobId jobId, ResultSetId rsId, int partition, int nPartitions,
             IResultMetadata metadata, boolean emptyResult) throws HyracksException;
 
-    void reportPartitionWriteCompletion(JobId jobId, ResultSetId resultSetId, int partition, int resultCount)
+    void reportPartitionWriteCompletion(JobId jobId, ResultSetId resultSetId, int partition, long resultCount)
             throws HyracksException;
 
     void reportPartitionConsumed(JobId jobId, ResultSetId rsId, int partition) throws HyracksException;
